server/src/services/vibe_matcher/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
deprecation==2.1.0
exceptiongroup==1.2.2
fastapi==0.109.2  # Added FastAPI
frozenlist==1.6.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.3
numpy==1.26.4  # Added NumPy explicitly
packaging==25.0
pillow==10.2.0  # Added Pillow for image processing
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9  # Added for handling form data
realtime==2.4.2
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tomli==2.2.1
transformers==4.37.2  # Added for CLIP model
typing-inspection==0.4.0
typing_extensions==4.13.2
websockets==14.2
yarl==1.20.0
torch==2.7.0
uvicorn==0.34.2

